Split query batch runner for #3380
