+++ changelog.d/unreleased/3380.internal.md
@@ -1,0 +1,16 @@
+---
+category: internal
+issues:
+  - 3380
+affected:
+  - src/CodeIndex/Cli/QueryCommandRunner.cs
+  - src/CodeIndex/Cli/QueryCommandRunner.Batch.cs
+---
+
+## English
+
+- **Split query batch orchestration into a focused partial (#3380)** — `QueryCommandRunner` now keeps batch-mode parsing, DB setup, and query dispatch in a dedicated partial without changing CLI behavior.
+
+## 日本語
+
+- **query batch orchestration を focused partial に分割しました (#3380)** — `QueryCommandRunner` の batch mode parsing、DB setup、query dispatch を専用 partial に移し、CLI 挙動は変更していません。
